+++ src/app/api/google/google.api.ts
@@ -1,5 +1,5 @@
 import axios from "axios"
-const googleApiKey = process.env.NEXT_PUBLIC_GOOGLE_API_KEY || undefined;
+const googleApiKey = process.env.NEXT_PUBLIC_GOOGLE_API_KEY;
 
 export async function validateAddressZipCode(address: string) {
   const apiUrl = `https://maps.googleapis.com/maps/api/geocode/json?address=${encodeURIComponent(
